requirements.txt
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
python-telegram-bot==13.0
